requirements.txt
markdown==3.7
Jinja2==3.1.5
PyYAML==6.0.2
Pygments==2.19.1
